+++ mini-projects/projects/day06-project-tuples2.py
@@ -1,0 +1,11 @@
+employees = [
+    {"name": "Adeeb", "department": "IT", "salary": 50000},
+    {"name": "Sara", "department": "HR", "salary": 45000},
+    {"name": "Ali", "department": "Sales", "salary": 40000}
+]
+
+for employee in employees:
+    print("Name:", employee["name"])
+    print("Department:", employee["department"])
+    print("Salary:", employee["salary"])
+    print()